Refactor Workflow tests from redundant class
